chore!: move old particle functions to deprecated
