+++ tests/canonical_formalism/test_ls_coupling.py
@@ -2,7 +2,7 @@ import pytest
 
 from expertsystem.state import particle
 from expertsystem.state.conservationrules import ParityConservationHelicity
-from expertsystem.state.particle import (
+from expertsystem.state.particle.deprecated import (
     InteractionQuantumNumberNames,
     Spin,
     SpinQNConverter,
@@ -131,7 +131,7 @@ def test_canonical_clebsch_gordan_ls_coupling(
 
     l_label = InteractionQuantumNumberNames.L
     s_label = InteractionQuantumNumberNames.S
-    qn_label = particle.LABELS.QuantumNumber
+    qn_label = particle.deprecated.LABELS.QuantumNumber
 
     spin_converter = SpinQNConverter()
     node_props = {
