+++ tests/channels/test_y_to_d0d0barpi0pi0.py
@@ -10,7 +10,7 @@ from expertsystem.amplitude.canonicaldecay import (
     CanonicalAmplitudeGeneratorXML,
 )
 from expertsystem.amplitude.helicitydecay import HelicityAmplitudeGeneratorXML
-from expertsystem.state.particle import (
+from expertsystem.state.particle.deprecated import (
     InteractionQuantumNumberNames,
     create_spin_domain,
 )
