+++ tests/conservation_laws/test_mass.py
@@ -1,5 +1,5 @@
 from expertsystem.state.conservationrules import MassConservation
-from expertsystem.state.particle import (
+from expertsystem.state.particle.deprecated import (
     ParticleDecayPropertyNames,
     ParticlePropertyNames,
 )
