+++ tox.ini
@@ -75,6 +75,8 @@ ignore_missing_imports = True
 ignore_errors = True
 [mypy-expertsystem.state.particle.*]
 ignore_errors = False
+[mypy-expertsystem.state.particle.deprecated.*]
+ignore_errors = True
 
 [mypy-tests.*]
 disallow_untyped_defs = False
